requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.3.3
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
joblib==1.3.2
magic-filter==1.0.12
multidict==6.0.5
numpy==1.21.6
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2024.2
scikit-learn==1.0.2
scipy==1.7.3
SQLAlchemy==2.0.36
threadpoolctl==3.1.0
typing_extensions==4.7.1
yarl==1.9.4
zipp==3.15.0
